tests: move commit_transactions() helper to testutils
